pyproject.toml
[project]
name = "helm-charts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "ruff>=0.15.11",
    "pytest",
    "meilisearch",
    "rich",
    "kubernetes",
    "ruamel.yaml",
    "pytest-snapshot",
    "pytest-cov",
    "pytest-sugar",
    "pytest-xdist",
    "packaging",
]

[tool.uv.workspace]
members = ["charts/meilisearch/resources"]
